Разбил проект на модули, внес зависимости между модулями
